requirements-serving.txt
# Serving dependencies
fastapi==0.103.1
uvicorn[standard]==0.23.2
scikit-learn==1.3.0
numpy==1.24.3
google-cloud-storage==2.10.0
pydantic==2.4.2
